refactor: update imports to use Assessable namespace for ChoiceAnswer components
