+++ src/components/documents/ChoiceAnswer/Hints/index.tsx
@@ -2,7 +2,7 @@ import { mdiCloseCircleOutline, mdiInformationOutline } from '@mdi/js';
 import Icon from '@mdi/react';
 import Button from '@tdev-components/shared/Button';
 import Card from '@tdev-components/shared/Card';
-import ChoiceAnswerDocument from '@tdev-models/documents/Quiz/ChoiceAnswer';
+import ChoiceAnswerDocument from '@tdev-models/documents/Assessable/ChoiceAnswer';
 import clsx from 'clsx';
 import { observer } from 'mobx-react-lite';
 import React from 'react';
