+++ src/models/documents/Assessable/iAssessable.ts
